Avoid ulimit on macOS
Bash's ulimit sets both the current and maximum limits, so that the
limit cannot be increased later. (macOS zsh only sets the current
limit.)

Removing this line means that the default limit for processes per user
will not be limited (the current OS default is much larger). The limit
for number of open files will be limited to 256 (rlim_cur) but can be
increased later by programs that call setrlimit() or ad hoc by ulimit.